extract ContainerHomeWebPreview component
